# perf(gatsby): more efficient parent-child check through arrays