@pvdz 2020-03-10
Seems to save ~30s in total on the ifsc benchmark: https://gist.github.com/pvdz/b4eeecb062ed28a48a6d7cc16b644de9